Make the Last-Modified header distinct
I bumped into a problem in the wild where the feed's "updated" property
was used to produce the If-Modified-Since header value.

This would apper to work if that timestamp and the Last-Modified
timestamp were the same but would likely cause troubles down the road.

Change the header's value to one distinct from the feed's contents to
make this kind of behavior more visible.